Crate draught

Draught

An implementation of checkers/draughts in Rust WebAssembly with a minimax AI player

Re-exports

pub use board::Board;

pub use game::Game;

pub use comp::Computer;

pub use paint::Painter;

Modules

board

Board module for components related to the checkers board and game structure

comp

AI player logic

game

Top-level object for managing Boards, applying and managing turns

paint

Components for painting board states onto HTML canvases

utils

Macros

log

Wrap the web_sys access to the browser console in a macro for easy logging

log_error

Constants

__wasm_bindgen_generated_init_wasm__const 🔒

Functions

init_wasm