Module draught::board

Board module for components related to the checkers board and game structure

Modules

enums

iter

Structs

Board

Single state of a checkers board

BrdIdx

Rank 2 tensor index to identify a board square by row and column

Direction

Standard diagonal movements given by north west/east etc

Piece

Game piece given by its team and strength (normal or kinged)

Square

Board squares given by a state and a possible occupying game piece

Constants

STD_HEIGHT

Standard height of a checkers board is 8 squares

STD_WIDTH

Standard width of a checkers board is 8 squares

__wbg_get_board_current_turn__const

__wbg_get_board_height__const

__wbg_get_board_width__const

__wbg_get_brdidx_col__const

__wbg_get_brdidx_row__const

__wbg_get_piece_strength__const

__wbg_get_piece_team__const

__wbg_get_square_occupant__const

__wbg_get_square_state__const