Struct draught::board::Board

pub struct Board {
    cells: Vec<Square>,
    pub width: usize,
    pub height: usize,
    pub current_turn: Team,
}

Single state of a checkers board

Fields
cells: Vec<Square>

1D backing array of board squares for the 2D game board

width: usizeheight: usizecurrent_turn: Team
Implementations

impl Board

pub fn cell_mut(&mut self, idx: usize) -> &mut Square

Get a mutable reference to a board square by 1D array index

pub fn adjacent_indices(&self, idx: BrdIdx) -> Option<Vec<usize>>

Get the 1D array indices for the diagonally adjacent squares of a given board square

Returns

None: If the given square is unplayable

Some(Vec): A variable length vector of 1D indices for diagonally adjacent squares. Vector may be between 1 and 4 items long depending on the location of the given square

pub fn adjacent_dir(&self, idx: BrdIdx) -> Option<Direction<Square>>

Get a Direction structure for the diagonally adjacent squares of a given board square

Similar to the Board::adjacent_indices function with differently formatted results

Returns

None: If the given square is unplayable

Some(Direction): A Direction structure for the diagonally adjacent squares.

pub fn filter_indices(
&self,
idx: BrdIdx,
player: Team,
indices: Vec<usize>
) -> Vec<usize>

Filter an array of diagonal indices (Like those from Board::adjacent_indices, Board::jumpable_indices) for those that are legal for a team’s man, i.e. solely up or down the board

pub fn player_adjacent_indices(
&self,
idx: BrdIdx,
player: Team
) -> Option<Vec<usize>>

pub fn jumpable_indices(&self, idx: BrdIdx) -> Option<Vec<usize>>

Get the 1D array indices for the diagonally jumpable squares of a given board square

Returns

None: If the given square is unplayable

Some(Vec): A variable length vector of 1D indices for diagonally jumpable squares. Vector may be between 1 and 4 items long depending on the location of the given square

pub fn jumpable_dir(&self, idx: BrdIdx) -> Option<Direction<Square>>

Get a Direction structure for the diagonally jumpable squares of a given board square

Similar to the Board::jumpable_indices function with differently formatted results

Returns

None: If the given square is unplayable

Some(Direction): A Direction structure for the diagonally jumpable squares.

pub fn player_jumpable_indices(
&self,
idx: BrdIdx,
player: Team
) -> Option<Vec<usize>>

pub fn idx_diffs(from: BrdIdx, to: BrdIdx) -> (isize, isize)

Get the difference between two BrdIdx objects

impl Board

pub fn cell(&self, idx: usize) -> Square

Get a copy of a board square by 1D array index

pub fn set_cell(&mut self, idx: usize, square: Square)

Get a copy of a board square by 1D array index

pub fn grid_cell(&self, idx: BrdIdx) -> Square

Get a copy of a board square by 2D BrdIdx index

pub fn cell_index(&self, row: usize, col: usize) -> usize

Transform a 2D row/column board index into a single 1D index for use with Board::cells

pub fn cell_idx(&self, idx: BrdIdx) -> usize

Similar to Board::cell_index but with a BrdIdx instead of separate indices. Transform a 2D row/column board index into a single 1D index for use with Board::cells

pub fn board_index(&self, idx: usize) -> BrdIdx

Transform a 1D array index (for Board::cells) into a 2D game board index (by row/col)

pub fn can_move(&self, from: BrdIdx, to: BrdIdx) -> Moveable

Check whether a move given by source and destination indices is legal

pub fn validate_man_move(
&self,
from: BrdIdx,
to: BrdIdx,
from_square_occupant: Piece
) -> Moveable

pub fn validate_king_move(
&self,
from: BrdIdx,
to: BrdIdx,
from_square_occupant: Piece
) -> Moveable

pub fn jumpee_idx(&self, from: BrdIdx, to: BrdIdx) -> usize

pub fn cells(&self) -> *const Square

Get a pointer to the backing array of board squares, Board::cells

pub fn num_cells(&self) -> usize

Get the number of board squares

pub fn num_pieces(&self) -> usize

Get the number of remaining pieces

pub fn num_player(&self, team: Team) -> usize

Get the number of remaining pieces for a player

pub fn score(&self) -> isize

Get the score value, Black - White pieces

pub fn cell_state(&self, idx: usize) -> SquareState

Get the state of a board square by 1D array index

pub fn apply_move(&self, from: BrdIdx, to: BrdIdx) -> Board

pub fn apply_jump(&self, from: BrdIdx, to: BrdIdx) -> Board

impl Board

pub fn validate_jumpee(jumpee: Square, from_occ: Piece) -> Moveable

Unwrap the jumpee piece from the square and Board::check_jumpee_team with Moveable response

pub fn check_jumpee_team(from: Piece, jumpee: Piece) -> bool

Check that the source piece and the jumpee are of opposing teams

pub fn new(width: usize, height: usize, current_turn: Team) -> Board

Initialise a game board without game pieces

pub fn init_game(board: Board, piece_rows: usize) -> Board

Reset the given board to a starting layout with 3 rows of opposing pieces