Struct draught::game::Game

pub struct Game {
    current: Board,
    selected_piece: Option<BrdIdx>,
    previous_boards: Vec<Board>,
    painter: Option<Painter>,
    search_depth: usize,
    pub last_node_count: usize,
    pub perfect_chance: f64,
}

Root-level structure for managing the game as a collection of board states

Fields
current: Boardselected_piece: Option<BrdIdx>previous_boards: Vec<Board>painter: Option<Painter>search_depth: usizelast_node_count: usizeperfect_chance: f64
Implementations

impl Game

pub fn previous_board(&self, turn: usize) -> &Board

Get a read-only copy of a previous turn’s board

pub fn current_board(&self) -> &Board

Set current board to given

impl Game

pub fn current_board_cells(&self) -> *const Square

Get pointer to current board’s squares

pub fn current_board_len(&self) -> usize

Get pointer to current board’s squares

pub fn current_turn(&self) -> Team

Current turn’s team

pub fn score(&self) -> isize

Current board’s score

pub fn winning(&self) -> Option<Team>

Get currently winning player

pub fn has_won(&self) -> Option<Team>

Check if a player has won

pub fn current_cell_state(&self, idx: &BrdIdx) -> Square

Get square on current board for given index

pub fn set_search_depth(&mut self, search_depth: usize)

Set tree depth for AI to search to

pub fn set_selected(&mut self, idx: &BrdIdx)

Set given index as selected piece TODO: Check whether valid square?

pub fn set_perfect_chance(&mut self, new_chance: f64)

Set proportion of perfect moves from AI

pub fn clear_selected(&mut self)

Clear currently selected piece

pub fn make_move(&mut self, from: BrdIdx, to: BrdIdx) -> Moveable

Attempt to make a move given a source and destination index

pub fn execute_move(&mut self, from: BrdIdx, to: BrdIdx)

Update board state with given move and push new board into current state

pub fn execute_jump(&mut self, from: BrdIdx, to: BrdIdx)

Update board state with given jump move and push new board into current state

pub fn push_new_board(&mut self, board: Board)

Push current board into the previous turns and set given board to current

pub fn set_current(&mut self, board: Board)

Set current board to given

pub fn new(
width: usize,
height: usize,
piece_rows: usize,
first_turn: Team,
search_depth: usize
) -> Game

Get new game without board renderer

pub fn new_with_canvas(
width: usize,
height: usize,
piece_rows: usize,
first_turn: Team,
search_depth: usize,
canvas_id: &str,
canvas_width: u32,
canvas_height: u32
) -> Game

Get a new game with canvas ID and dimensions

pub fn set_painter(&mut self, value: Painter)

Set painter for rendering boards

pub fn draw(&self)

Draw current board using painter if exists

pub fn ai_move(&mut self)

Create computer, get move from current board and update current board