Module draught::paint

Components for painting board states onto HTML canvases

Structs

Painter

Used to paint boards onto HTML canvases

Constants

BLACK_PIECE 🔒

Default hex colour value for black pieces

BLACK_PIECE_OUTLINE 🔒

Default hex colour value for black piece outline

BLACK_SQUARE 🔒

Default hex colour value for black square background

DRAW_OUTLINE 🔒

Whether to outline black squares

DRAW_PIECE_OUTLINES 🔒

Whether to outline pieces

KING_OUTLINE 🔒

Default hex colour value for black piece outline

OUTLINE_WIDTH 🔒

Line width when outlining black squares as proportion of min cell dimension

PIECE_OUTLINE_PROPORTION 🔒

Line width for outlining pieces as proportion of piece radius

PIECE_PROPORTION 🔒

Proportion of square that piece fills as proportion of min cell dimension

SELECTED_PIECE 🔒

Default hex colour value for selected piece

SELECTED_PIECE_OUTLINE 🔒

Default hex colour value for selected piece outline

SQUARE_OUTLINE 🔒

Default hex colour value for outline of black squares

WHITE_PIECE 🔒

Default hex colour value for white pieces

WHITE_PIECE_OUTLINE 🔒

Default hex colour value for white piece outline

WHITE_SQUARE 🔒

Default hex colour value for white square background