Struct draught::paint::Painter

pub struct Painter {

    canvas: HtmlCanvasElement,
    context: CanvasRenderingContext2d,
    selected_idx: Option<BrdIdx>,
    white_square: JsValue,
    black_square: JsValue,
    white_piece: JsValue,
    black_piece: JsValue,
    selected_piece: JsValue,
    white_piece_line: JsValue,
    black_piece_line: JsValue,
    selected_piece_line: JsValue,
    king_line: JsValue,
    piece_lines: bool,
    piece_line_proportion: f64,
    square_outline: JsValue,
    outline_width: f64,
    draw_outline: bool,
    width: u32,
    height: u32,

}

Used to paint boards onto HTML canvases

Fields
canvas: HtmlCanvasElementcontext: CanvasRenderingContext2dselected_idx: Option<BrdIdx>white_square: JsValueblack_square: JsValuewhite_piece: JsValueblack_piece: JsValueselected_piece: JsValuewhite_piece_line: JsValueblack_piece_line: JsValueselected_piece_line: JsValueking_line: JsValuepiece_lines: boolpiece_line_proportion: f64square_outline: JsValueoutline_width: f64draw_outline: boolwidth: u32height: u32
Implementations

impl Painter

pub fn set_selected(&mut self, idx: &Option<BrdIdx>)

Set selected piece by board index

fn get_canvas(canvas_id: &str) -> HtmlCanvasElement

Get a canvas by element ID

fn get_canvas_context(canvas: &HtmlCanvasElement) -> CanvasRenderingContext2d

Get a 2D canvas context for a given canvas

impl Painter

pub fn new(width: u32, height: u32, canvas_id: &str) -> Painter

Default constructor which queries for canvas by ID

pub fn new_with_canvas(
width: u32,
height: u32,
canvas: HtmlCanvasElement
) -> Painter

Constructor with given canvas element

pub fn set_square_outline(&mut self, value: JsValue)

Set new square outline colour value

pub fn set_outline_width(&mut self, value: f64)

Set new line width for outlining squares

pub fn set_draw_outline(&mut self, value: bool)

Set whether squares are outlined

pub fn reset_dimensions(&self)

Reset the canvas dimensions to the given width and height

pub fn validate_board_dim(&self, board: &Board) -> bool

Check whether given canvas dimensions divide evenly by given board dimenions

pub fn draw(&self, board: &Board)

Draw a board onto the canvas